ganarcrm_django/requirements.txt
django
djangorestframework
djoser
django-cors-headers
dj-stripe